refactor(integer): split APKINDEX validation helpers
